dev 3 fix (Мелкие правки тестов)
